requirements.txt
psycopg2-binary
dj-database-url
Pillow
django-q2